requirements.txt
# Core Framework Layer (Strict Lock)
fastapi==0.115.11  # Core API framework needs absolute stability
uvicorn==0.19.0    # ASGI server tightly coupled with FastAPI
pydantic==2.6.1    # Core data validation library

# AI Integration Layer (Compatible Lock)
langchain>=0.1.0,<0.2.0   # Allow patch updates but maintain major version
openai==0.28
sentence-transformers~=2.6.0  # Allow 2.6.x security updates

# Database Layer (Security Updates)
pgvector>=0.4.0          # Vector extension needs strict PG version match
psycopg2-binary>=2.9.9   # Allow security patch updates

# Document Processing Layer (Feature Updates)
unstructured>=0.17.0,<0.18.0  # Document parsing library under active development
pymupdf~=1.23.0          # Maintain PDF rendering compatibility
python-docx==1.1.0       # Mature library, no frequent updates needed
openpyxl>=3.1.5          # Upgrade for Excel new features

# Tool Chain Layer (Flexible Updates)
python-dotenv>=1.0.0     # Configuration tool allows optimization
langchain-community>=0.0.6  # Community plugins need new features

# UI Layer
gradio>=5.21.0